+++ services/platform/operator/controller/apps/model.go
@@ -1,4 +1,4 @@
-package controller
+package apps
 
 type (
 	AppValues struct {
