+++ services/platform/operator/controller/apps/secrets.go
@@ -1,4 +1,4 @@
-package controller
+package apps
 
 import (
 	"context"
